Ensure a full parse when getting the spine parser
If we're already going to traverse the entire snapshot, might as well do a proper parse as it's most likely already requested by other services or will be requested soon anyway. Fast forwarding a throw-away spine parser is wasted effort. Better pay a bit more once but have a parse tree ready for everyone else to use.

If we don't do this, we fall through to "XML parser failed to recover any state" anyway.

This has been shipping in VSMac and got some coverage.