Use custom editor logger factory
This means we can compose without worrying that we might conflict with
a generic ILogger<T> impl in the composition that another extension or
the host might wire up.

It also means we don't have to export every logger individually, and
allows the factory to provide buffer/editor specific loggers
if desired in future.