Alter how XmlCompletionSource gathers items
The list can now contain items of multiple kinds, such as
element values and child elements values.

When there are multiple kinds, the tasks will run in parallel.

There is a virtual method that allows subclasses to add additional
items to the list, and this will also run in parallel with the
other tasks.