Remove MiniEditor dep and update VS editor packages
VS Editor added an internal Disconnect() member to the IConnectedSpan
interface in 17.6, which meant that MiniEditor could not implement
any version of the VS Editor APIs newer than 17.5.

Not being able to update the VS Editor packages caused issues for
MSBuildEditor, especially when it needed to bring in newer versions
of Roslyn for the LSP.

Fix this by removing MiniEditor and instead using the VS Editor itself
to run the tests, mimicking how Roslyn does it. This has the downside
that we can no longer run the editor tests on Mac. However, as Visual
Studio for Mac is now discontinued, this is no longer a pressing concern.